Add redis password parameter
Redis server API was not protected with a password
by default.
To fix this issue we need to provide a way to
configure redis with authentification in salt-formula.
This patch addresses the problem for redis used
by contrail analytics.
If redis password is not set in the model then redis
will be configured without password.

Co-Authored-By Gleb Zimin <gzimin@mirantis.com>
Co-Authored-By Pavel Svimbersky <psvimbersky@mirantis.com>

Change-Id: I05813169be38ece1bd592d73a3ba1f4c7502d602
Related-Prod: PROD-22096